feat: load thedapps data in advance (on the Asset page)
