+++ src/components/dapp-staking/stake-manage/StakeManage.vue
@@ -41,25 +41,25 @@
 <script lang="ts">
 import MobileNavigator from 'src/components/assets/transfer/MobileNavigator.vue';
 import BackToPage from 'src/components/common/BackToPage.vue';
-import StakeForm from 'src/components/dapp-staking/stake-manage/StakeForm.vue';
-import SelectFunds from 'src/components/dapp-staking/stake-manage/SelectFunds.vue';
-import StakeInformation from 'src/components/dapp-staking/stake-manage/StakeInformation.vue';
 import ModalSelectFunds from 'src/components/dapp-staking/stake-manage/ModalSelectFunds.vue';
+import SelectFunds from 'src/components/dapp-staking/stake-manage/SelectFunds.vue';
+import StakeForm from 'src/components/dapp-staking/stake-manage/StakeForm.vue';
+import StakeInformation from 'src/components/dapp-staking/stake-manage/StakeInformation.vue';
+import { WalletModalOption } from 'src/config/wallets';
 import {
   useBreakpoints,
-  useNetworkInfo,
+  useDappRedirect,
+  useDispatchGetDapps,
   useStake,
   useStakingList,
-  useDappRedirect,
 } from 'src/hooks';
 import { wait } from 'src/hooks/helper/common';
 import { Path } from 'src/router';
 import { useStore } from 'src/store';
-import { computed, defineComponent, ref, watchEffect, watch } from 'vue';
-import { useRoute } from 'vue-router';
 import { DappCombinedInfo } from 'src/v2/models';
-import { WalletModalOption } from 'src/config/wallets';
+import { computed, defineComponent, ref, watch } from 'vue';
 import { useI18n } from 'vue-i18n';
+import { useRoute } from 'vue-router';
 
 export type StakeRightUi = 'information' | 'select-funds-from';
 
@@ -78,9 +78,9 @@ export default defineComponent({
 
     const { t } = useI18n();
     const { screenSize, width } = useBreakpoints();
-    const { currentNetworkName } = useNetworkInfo();
     const route = useRoute();
     useDappRedirect();
+    useDispatchGetDapps();
     const { setAddressTransferFrom, formattedTransferFrom, currentAccount, handleStake } =
       useStake();
 
@@ -103,17 +103,6 @@ export default defineComponent({
         if (ui === 'select-funds-from') {
           isModalSelectFunds.value = true;
         }
-      }
-    };
-
-    const dispatchGetDapps = (): void => {
-      const isDispatch =
-        currentNetworkName.value && dapps.value.length === 0 && currentAccount.value;
-      if (isDispatch) {
-        store.dispatch('dapps/getDapps', {
-          network: currentNetworkName.value.toLowerCase(),
-          currentAccount: currentAccount.value,
-        });
       }
     };
 
@@ -143,8 +132,6 @@ export default defineComponent({
       await setRightUi('information');
       isModalSelectFunds.value && handleModalSelectFunds({ isOpen: false });
     };
-
-    watchEffect(dispatchGetDapps);
 
     watch(
       [currentAccount, dapp],
